rename providers to repositories
* use repositories as raw data sources and providers and domain model sources
* removed uninitialized (unused)
