+++ balancer-js/src/index.ts
@@ -13,9 +13,7 @@ export * from './modules/swaps/swaps.module';
 export * from './modules/subgraph/subgraph.module';
 export * from './modules/sor/sor.module';
 export * from './modules/pools/pools.module';
-export * from './modules/data-providers/pool/static.provider';
-export * from './modules/data-providers/token/static.provider';
-export * from './modules/data-providers/token-price/static.provider';
+export * from './modules/data';
 export * from './balancerErrors';
 export {
   SwapInfo,
