+++ balancer-js/src/modules/data/index.ts
@@ -1,0 +1,3 @@
+export * from './pool';
+export * from './token';
+export * from './token-price';
