+++ balancer-js/src/modules/data/pool/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './fallback';
+export * from './static';
