+++ balancer-js/src/modules/data/token-price/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './static';
