+++ balancer-js/src/modules/data/token-price/static.ts
@@ -1,7 +1,7 @@
 import { BigNumber, formatFixed } from '@ethersproject/bignumber';
 import { parseFixed } from '@/lib/utils/math';
 import { Price, TokenPrices } from '@/types';
-import { TokenPriceProvider } from './provider.interface';
+import { TokenPriceProvider } from './types';
 
 const SCALING_FACTOR = 18;
