+++ balancer-js/src/modules/data/token/index.ts
@@ -1,0 +1,2 @@
+export * from './static';
+export * from './types';
